split header related functions to HeaderManager
