+++ src/MarkdownTocTools.ts
@@ -2,15 +2,15 @@ import {
     window,
     Position,
     Range,
-    TextEditorEdit,
-    TextDocument
-} from 'vscode';
+    TextEditorEdit} from 'vscode';
 import { Header } from "./models/Header";
-import { ConfigManager } from './configManager';
+import { ConfigManager } from './ConfigManager';
+import { HeaderManager } from './HeaderManager';
 
 export class MarkdownTocTools {
 
     configManager = new ConfigManager();
+    headerManager = new HeaderManager();
 
     // Public function
     public updateMarkdownToc(isBySave: boolean = false) {
@@ -26,8 +26,10 @@ export class MarkdownTocTools {
         editor.edit(function (editBuilder) {
             let tocRange = markdownTocTools.getTocRange();
             markdownTocTools.updateOptions(tocRange);
-            if (isBySave && ((!markdownTocTools.configManager.options.UPDATE_ON_SAVE) || (tocRange == null)))
-                return false;
+            if (isBySave && ((!markdownTocTools.configManager.options.UPDATE_ON_SAVE) || (tocRange == null))) {
+                return false
+            };
+
             // save options, and delete last insert
             if (tocRange != null) {
                 insertPosition = tocRange.start;
@@ -35,7 +37,7 @@ export class MarkdownTocTools {
                 markdownTocTools.deleteAnchor(editBuilder);
             }
 
-            let headerList = markdownTocTools.getHeaderList();
+            let headerList = markdownTocTools.headerManager.getHeaderList();
 
             markdownTocTools.createToc(editBuilder, headerList, insertPosition);
             markdownTocTools.insertAnchor(editBuilder, headerList);
@@ -62,7 +64,7 @@ export class MarkdownTocTools {
     public updateMarkdownSections() {
         let tocRange = this.getTocRange();
         this.updateOptions(tocRange);
-        let headerList = this.getHeaderList();
+        let headerList = this.headerManager.getHeaderList();
         let editor = window.activeTextEditor;
         if (editor != undefined) {
             editor.edit(function (editBuilder) {
@@ -76,7 +78,7 @@ export class MarkdownTocTools {
     public deleteMarkdownSections() {
         let tocRange = this.getTocRange();
         this.updateOptions(tocRange);
-        let headerList = this.getHeaderList();
+        let headerList = this.headerManager.getHeaderList();
         let editor = window.activeTextEditor;
         if (editor != undefined && headerList != undefined) {
             editor.edit(function (editBuilder) {
@@ -227,104 +229,6 @@ export class MarkdownTocTools {
 
     private getIndiceOfDepth() {
         return Array.apply(null, new Array(this.configManager.options.DEPTH_TO.value - this.configManager.options.DEPTH_FROM.value + 1)).map(Number.prototype.valueOf, 0);
-    }
-
-    private getHeaderList() {
-        let headerList: Header[] = [];
-        let editor = window.activeTextEditor;
-
-        if (editor != undefined) {
-            let doc = editor.document;
-
-            for (let index = 0; index < doc.lineCount; index++) {
-                let lineText = this.getNextLineIsNotInCode(index, doc);
-
-                let header = this.getHeader(lineText);
-
-                if (header.isHeader) {
-                    header.orderArray = this.calculateHeaderOrder(header, headerList);
-                    header.range = new Range(index, 0, index, lineText.length);
-                    headerList.push(header);
-                }
-            }
-        }
-
-        return headerList;
-    }
-
-    private getNextLineIsNotInCode(index: number, doc: TextDocument) {
-        let lineText = doc.lineAt(index).text;
-
-        let isCodeStyle1 = lineText.match(this.configManager.optionKeys.REGEXP_CODE_BLOCK1) != null;
-        let isCodeStyle2 = lineText.match(this.configManager.optionKeys.REGEXP_CODE_BLOCK2) != null;
-
-        let nextIndex = index;
-
-        while (isCodeStyle1 || isCodeStyle2) {
-            nextIndex = index + 1;
-
-            let nextLine = doc.lineAt(nextIndex).text;
-
-            isCodeStyle1 = nextLine.match(this.configManager.optionKeys.REGEXP_CODE_BLOCK1) != null;
-            isCodeStyle2 = nextLine.match(this.configManager.optionKeys.REGEXP_CODE_BLOCK2) != null;
-        }
-
-        return doc.lineAt(nextIndex).text;
-    }
-
-    private getHeader(lineText: string) {
-        let header = new Header(this.configManager.options.ANCHOR_MODE.value);
-
-        let headerTextSplit = lineText.match(this.configManager.optionKeys.REGEXP_HEADER_META);
-
-        if (headerTextSplit != null) {
-            header.headerMark = headerTextSplit[1];
-            header.orderedListString = headerTextSplit[2];
-            header.dirtyTitle = headerTextSplit[4];
-        }
-
-        return header;
-    }
-
-    private calculateHeaderOrder(headerBeforePushToList: Header, headerList: Header[]) {
-
-        if (headerList.length == 0) {
-            // special case: First header with depth = 1
-            return [1];
-        }
-
-        let lastheaderInList = headerList[headerList.length - 1];
-
-        if (headerBeforePushToList.depth < lastheaderInList.depth) {
-            // continue of the parent level
-            let previousheader = headerList.find(header => header.depth == headerBeforePushToList.depth);
-
-            if (previousheader != undefined) {
-                let orderArray = Object.assign([], previousheader.orderArray);
-                orderArray[orderArray.length - 1]++;
-
-                return orderArray;
-            }
-        }
-
-        if (headerBeforePushToList.depth > lastheaderInList.depth) {
-            // child level of previous
-            // order start with 1
-            let orderArray = Object.assign([], lastheaderInList.orderArray);
-            orderArray.push(1);
-
-            return orderArray;
-        }
-
-        if (headerBeforePushToList.depth == lastheaderInList.depth) {
-            // the same level, increase last item in orderArray
-            let orderArray = Object.assign([], lastheaderInList.orderArray);
-            orderArray[orderArray.length - 1]++;
-
-            return orderArray;
-        }
-
-        return [];
     }
 
     dispose() {
